requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
dj-database-url==3.0.1
python-dotenv==1.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1 